allow user to add repeat notification on days

+++ TrivaAlarm/AppDelegate.swift
@@ -19,22 +19,19 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
         application.cancelAllLocalNotifications()
-       let defualts = NSUserDefaults.standardUserDefaults()
-        defualts.setValue(22.5, forKey: "boss")
-
-
-
-        let rootViewController = self.window?.rootViewController as! UINavigationController
-
-        let viewController = rootViewController.topViewController as! ViewController
-        if viewController.respondsToSelector("setCoreDataStack:"){
+        guard let rootViewController = self.window?.rootViewController as? UINavigationController else {
+            return false
+        }
+        guard let viewController = rootViewController.topViewController as? ViewController else {
+            return false
+        }
+        if viewController.respondsToSelector("setCoreDataStack:") {
             viewController.performSelector("setCoreDataStack:", withObject: coreDataStack)
         }
 
-
         UINavigationBar.appearance().barTintColor = UIColor(red: (49/255), green: (128/255), blue: (197/255), alpha: 1)
         UINavigationBar.appearance().tintColor = UIColor.whiteColor()
-        UINavigationBar.appearance().titleTextAttributes = [NSForegroundColorAttributeName : UIColor.whiteColor()]
+        UINavigationBar.appearance().titleTextAttributes = [NSForegroundColorAttributeName: UIColor.whiteColor()]
 
         let defaults = NSUserDefaults.standardUserDefaults()
         let isPreloaded = defaults.boolForKey("isPreloaded")
@@ -49,103 +46,117 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             defaults.setBool(true, forKey: "firstLoad")
         }
 
-
         UINavigationBar.appearance().barStyle = .Black
-
 
         application.registerUserNotificationSettings(UIUserNotificationSettings(forTypes: [UIUserNotificationType.Sound, UIUserNotificationType.Alert, UIUserNotificationType.Badge], categories: nil))
 
         if let options = launchOptions {
 
             if let notification = options[UIApplicationLaunchOptionsLocalNotificationKey] as? UILocalNotification {
-
-
                       dispatch_async(dispatch_get_main_queue(), {
-
-
                         let storyboard = UIStoryboard(name: "Main", bundle: nil)
-                        let vc = storyboard.instantiateViewControllerWithIdentifier("QuestionVC") as! QuestionViewController
+                        guard let vc = storyboard.instantiateViewControllerWithIdentifier("QuestionVC") as? QuestionViewController else {
+                            return
+                        }
 
                         vc.coreDataStack = self.coreDataStack
                         self.window?.rootViewController?.presentViewController(vc, animated: true, completion: nil)
-
-
-                        // NSNotificationCenter.defaultCenter().postNotificationName("questionSegue", object: nil)
-                        let numbOfQuestion:AnyObject = notification.userInfo!["NumberOfQuestion"]!
-                        let typeOfQuestion:AnyObject = notification.userInfo!["TypeOfQuestion"]!
-                        let alarmDate:AnyObject = notification.userInfo!["AlarmDate"]!
-                        let alarmSound:AnyObject = notification.userInfo!["AlarmSound"]!
+                        guard let notificationUserInfo = notification.userInfo  else {
+                            return
+                        }
+                        guard let numbOfQuestion: AnyObject = notificationUserInfo["NumberOfQuestion"] else {
+                            return
+                        }
+                        guard let typeOfQuestion: AnyObject = notification.userInfo!["TypeOfQuestion"] else {
+                            return
+                        }
+                        guard let alarmDate: AnyObject = notification.userInfo!["AlarmDate"] else {
+                            return
+                        }
+                        guard let alarmSound: AnyObject = notification.userInfo!["AlarmSound"] else {
+                            return
+                        }
                         let nc = NSNotificationCenter.defaultCenter()
-                        nc.postNotificationName("localNotficaionUserInfo", object: nil, userInfo: ["NumberOfQuestion":numbOfQuestion , "TypeOfQuestion": typeOfQuestion, "AlarmDate": alarmDate, "AlarmSound":alarmSound])
+                        nc.postNotificationName("localNotficaionUserInfo", object: nil, userInfo: ["NumberOfQuestion": numbOfQuestion , "TypeOfQuestion": typeOfQuestion, "AlarmDate": alarmDate, "AlarmSound": alarmSound])
                         application.cancelLocalNotification(notification)
 
                 })
 
                  application.cancelLocalNotification(notification)
-
-
-
             }
         }
-
 
         return true
     }
 
     func createFirstAlarm() {
-        let entity = NSEntityDescription.entityForName("Alarms", inManagedObjectContext:coreDataStack.managedObjectContext)
+        let entity = NSEntityDescription.entityForName("Alarms", inManagedObjectContext: coreDataStack.managedObjectContext)
         let alarm = NSManagedObject(entity: entity!, insertIntoManagedObjectContext: coreDataStack.managedObjectContext)
 
         alarm.setValue("Wake Up", forKey: "name")
         alarm.setValue(NSDate(), forKey: "time")
         alarm.setValue(1, forKey: "numOfQuestionsToEnd")
-        alarm.setValue(false, forKey: "snooze")
         alarm.setValue("Random", forKey: "questionType")
         alarm.setValue("LoudAlarm.wav", forKey: "alertSound")
         alarm.setValue(false, forKey: "on")
 
         coreDataStack.saveMainContext()
     }
-  
-        func application(application: UIApplication, didReceiveLocalNotification notification: UILocalNotification) {
+
+    func notficationRecieveSetup(notification: UILocalNotification) {
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        guard let vc = storyboard.instantiateViewControllerWithIdentifier("QuestionVC") as? QuestionViewController else {
+            return
+        }
+        vc.coreDataStack = self.coreDataStack
+        self.window?.rootViewController?.presentViewController(vc, animated: true, completion: nil)
+        guard let notificationUserInfo = notification.userInfo  else {
+            return
+        }
+        guard let numbOfQuestion: AnyObject = notificationUserInfo["NumberOfQuestion"] else {
+            return
+        }
+        guard let typeOfQuestion: AnyObject = notification.userInfo!["TypeOfQuestion"] else {
+            return
+        }
+        guard let alarmDate: AnyObject = notification.userInfo!["AlarmDate"] else {
+            return
+        }
+        guard let alarmSound: AnyObject = notification.userInfo!["AlarmSound"] else {
+            return
+        }
+        let nc = NSNotificationCenter.defaultCenter()
+        nc.postNotificationName("localNotficaionUserInfo", object: nil, userInfo: ["NumberOfQuestion": numbOfQuestion , "TypeOfQuestion": typeOfQuestion, "AlarmDate": alarmDate, "AlarmSound": alarmSound])
+
+    }
+
+    func application(application: UIApplication, didReceiveLocalNotification notification: UILocalNotification) {
             application.cancelAllLocalNotifications()
-
-                let storyboard = UIStoryboard(name: "Main", bundle: nil)
-                let vc = storyboard.instantiateViewControllerWithIdentifier("QuestionVC") as! QuestionViewController
-             vc.coreDataStack = self.coreDataStack
-                self.window?.rootViewController?.presentViewController(vc, animated: true, completion: nil)
-
-                let numbOfQuestion:AnyObject = notification.userInfo!["NumberOfQuestion"]!
-                let typeOfQuestion:AnyObject = notification.userInfo!["TypeOfQuestion"]!
-                let alarmDate:AnyObject = notification.userInfo!["AlarmDate"]!
-            let alarmSound:AnyObject = notification.userInfo!["AlarmSound"]!
-                let nc = NSNotificationCenter.defaultCenter()
-                nc.postNotificationName("localNotficaionUserInfo", object: nil, userInfo: ["NumberOfQuestion":numbOfQuestion , "TypeOfQuestion": typeOfQuestion, "AlarmDate": alarmDate, "AlarmSound":alarmSound])
-                application.cancelLocalNotification(notification)
+           notficationRecieveSetup(notification)
 
 
     }
 
-    func parseCSV (contentsOfURL: NSURL, encoding: NSStringEncoding) throws -> [(question:String, type:String, optionA: String, optionB: String, optionC:String, optionD:String, correctAnswer:String)] {
+    func parseCSV (contentsOfURL: NSURL, encoding: NSStringEncoding) throws -> [(question: String, type: String, optionA: String, optionB: String, optionC: String, optionD: String, correctAnswer: String)] {
         var error: NSError! = NSError(domain: "Migrator", code: 0, userInfo: nil)
         // Load the CSV file and parse it
         let delimiter = ","
-        var items:[(question:String, type:String, optionA: String, optionB: String, optionC:String, optionD:String, correctAnswer:String)]?
+        var items: [(question: String, type: String, optionA: String, optionB: String, optionC: String, optionD: String, correctAnswer: String)]?
 
         do {
             let content = try String(contentsOfURL: contentsOfURL, encoding: encoding)
             items = []
-            let lines:[String] = content.componentsSeparatedByCharactersInSet(NSCharacterSet.newlineCharacterSet()) as [String]
+            let lines: [String] = content.componentsSeparatedByCharactersInSet(NSCharacterSet.newlineCharacterSet()) as [String]
 
             for line in lines {
-                var values:[String] = []
+                var values: [String] = []
                 if line != "" {
                     // For a line with double quotes
                     // we use NSScanner to perform the parsing
                     if line.rangeOfString("\"") != nil {
-                        var textToScan:String = line
-                        var value:NSString?
-                        var textScanner:NSScanner = NSScanner(string: textToScan)
+                        var textToScan: String = line
+                        var value: NSString?
+                        var textScanner: NSScanner = NSScanner(string: textToScan)
                         while textScanner.string != "" {
 
                             if (textScanner.string as NSString).substringToIndex(1) == "\"" {
@@ -157,6 +168,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                             }
 
                             // Store the value into the values array
+
                             values.append(value as! String)
 
                             // Retrieve the unscanned remainder of the string
@@ -170,10 +182,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
                         // For a line without double quotes, we can simply separate the string
                         // by using the delimiter (e.g. comma)
-                    } else  {
+                    } else {
                         values = line.componentsSeparatedByString(delimiter)
                     }
-                    
                     // Put the values into the tuple and add it to the items array
                     let item = (question: values[0], type: values[1], optionA: values[2], optionB: values[3], optionC: values[4], optionD: values[5], correctAnswer: values[6])
                     items?.append(item)
@@ -182,16 +193,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         } catch let error1 as NSError {
             error = error1
         }
-        
         if let value = items {
             return value
         }
         throw error
     }
 
-    
-
-    func preloadData () {
+    func preloadData() {
         // Retrieve data from the source file
         if let contentsOfURL = NSBundle.mainBundle().URLForResource("questionCSV", withExtension: "csv") {
 
@@ -204,7 +212,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 // Preload the menu items
 
                     for item in items {
-                        let questionItem = NSEntityDescription.insertNewObjectForEntityForName("Questions", inManagedObjectContext: coreDataStack.managedObjectContext) as! Questions
+                        guard let questionItem = NSEntityDescription.insertNewObjectForEntityForName("Questions", inManagedObjectContext: coreDataStack.managedObjectContext) as? Questions else {
+                            return
+                        }
                         questionItem.question = item.question
                         questionItem.type = item.type
                         questionItem.optionA = item.optionA
@@ -212,9 +222,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                         questionItem.optionC = item.optionC
                         questionItem.optionD = item.optionD
                         questionItem.correctAnswer = item.correctAnswer
-
                         coreDataStack.saveMainContext()
-                    
                 }
 
             } catch let error1 as NSError {
@@ -225,10 +233,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 }
                 alertController.addAction(cancelAction)
 
-                let OKAction = UIAlertAction(title: "Ok", style: .Default) { (action) in
+                let oKAction = UIAlertAction(title: "Ok", style: .Default) { (action) in
                     // ...
                 }
-                alertController.addAction(OKAction)                                //Present the AlertController
+                alertController.addAction(oKAction)                                // Present the AlertController
                UIApplication.sharedApplication().keyWindow?.rootViewController?.presentViewController(alertController, animated: true, completion: nil)
             }
         }
@@ -240,7 +248,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             let fetchRequest = NSFetchRequest(entityName: "Questions")
 
             do{
-            let questionItems = (try! coreDataStack.managedObjectContext.executeFetchRequest(fetchRequest)) as! [Questions]
+                guard let questionItems = (try! coreDataStack.managedObjectContext.executeFetchRequest(fetchRequest)) as? [Questions] else {
+                    return
+                }
 
                 for question in questionItems {
                     coreDataStack.managedObjectContext.deleteObject(question)
@@ -257,10 +267,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 }
                 alertController.addAction(cancelAction)
 
-                let OKAction = UIAlertAction(title: "Ok", style: .Default) { (action) in
+                let oKAction = UIAlertAction(title: "Ok", style: .Default) { (action) in
                     // ...
                 }
-                alertController.addAction(OKAction)                                //Present the AlertController
+                alertController.addAction(oKAction)
+                // Present the AlertController
                 UIApplication.sharedApplication().keyWindow?.rootViewController?.presentViewController(alertController, animated: true, completion: nil)
 
             }
@@ -301,4 +312,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
 }
-
